+++ src/mediarithmics/api/plugin/audiencefeedconnector/AudienceFeedConnectorPluginResponseInterface.ts
@@ -95,6 +95,20 @@ export class MissingConfigurationPropertyError extends AudienceFeedInstanceConte
   }
 }
 
+export class MandatoryPropertyValueError extends AudienceFeedInstanceContextError {
+  public log: string;
+
+  constructor(
+    public feed_id: string,
+    public property_name: string,
+  ) {
+    super(
+      `${property_name} is a mandatory property. Please provide it when creating the feed.`,
+    );
+    this.log = `Mandatory property: ${property_name}`;
+  }
+}
+
 export class InvalidPropertyValueError extends AudienceFeedInstanceContextError {
   public log: string;