+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 # Unreleased
 
+- Add new Error class to have generic end user message: `MandatoryPropertyValueError`
+
 # 0.28.0 - 2024-07-09
 
 - Add new Error classes to have generic end user messages: `MissingConfigurationPropertyError`,